refactor player data
