+++ components/players/data.js
@@ -1,0 +1,29 @@
+const data = [
+    {
+      image: '/static/images/players/370/yun.jpg',
+      name: 'Yun',
+      desc: 'Boardman',
+      url: 'yun'
+    },
+    {
+      image: '/static/images/players/370/sean.jpg',
+      name: 'Sean',
+      desc: 'Space Jam',
+      url: 'sean'
+    },
+    {
+      image: '/static/images/players/370/russ.jpg',
+      name: 'Russ',
+      desc: 'King',
+      url: 'russ'
+    },
+    ,
+    {
+      image: '/static/images/players/370/george.jpg',
+      name: 'George',
+      desc: 'Big Shot from Downtown',
+      url: 'george'
+    }
+  ];
+
+  export default data;
