+++ components/players/index.js
@@ -1,33 +1,7 @@
-import Layout from '../layout';
 import Link from 'next/link';
+import Layout from '../layout';
+import data from './data';
 
-const data = [
-  {
-    image: '/static/images/players/370/yun.jpg',
-    name: 'Yun',
-    desc: 'Boardman',
-    url: 'yun'
-  },
-  {
-    image: '/static/images/players/370/sean.jpg',
-    name: 'Sean',
-    desc: 'Space Jam',
-    url: 'sean'
-  },
-  {
-    image: '/static/images/players/370/russ.jpg',
-    name: 'Russ',
-    desc: 'King',
-    url: 'russ'
-  },
-  ,
-  {
-    image: '/static/images/players/370/george.jpg',
-    name: 'George',
-    desc: 'Big Shot from Downtown',
-    url: 'george'
-  }
-];
 
 const players = () => {
   const items = data.map((d, i) => {
